% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5-functions.R
\name{h5Class}
\alias{h5Class}
\alias{is.H5D}
\alias{is.H5Group}
\title{Get the class of an HDF5 link}
\usage{
h5Class(file, name)

is.H5D(file, name)

is.H5Group(file, name)
}
\arguments{
\item{file}{An existing HDF5 file}

\item{name}{Name of a link in \code{file}}
}
\description{
Functions to get or check the class of an HDF5 link.
}
\examples{
file <- system.file("extdata", "pbmc_small.h5ad", package = "hdf5r.Extra")
h5Class(file, "X")
h5Class(file, "obs")
is.H5D(file, "X")
is.H5Group(file, "obs")

}
