% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/hdf5-methods.R
\name{h5CreateFile}
\alias{h5CreateFile}
\alias{h5CreateFile.character}
\title{Create a new HDF5 file}
\usage{
h5CreateFile(x, ...)

\method{h5CreateFile}{character}(x, ...)
}
\arguments{
\item{x}{Name of the new HDF5 file.}

\item{...}{Arguments passed to \code{H5File$new()}}
}
\value{
This is an operation function and no return. Any failure should raise an 
error.
}
\description{
A wrapper for \code{\link[hdf5r]{H5File}}\code{$new()}. If \code{file} 
exists, will only raise a warning.
}
\examples{
tmp.file <- tempfile(fileext = ".h5")
h5CreateFile(tmp.file)

}
