% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/hdf5-methods.R
\name{h5Prep}
\alias{h5Prep}
\alias{h5Prep.default}
\title{Prepare an R object to be written into HDF5 file}
\usage{
h5Prep(x, ...)

\method{h5Prep}{default}(x, ...)
}
\arguments{
\item{x}{The R object to be transformed}

\item{...}{Arguments to be passed to other methods}
}
\value{
An R object of converted \code{x}.
}
\description{
Methods to transform a complex R object (for example, S4 object) into 
combination of base R objects, such as \code{vector}, \code{array}, 
\code{data.frame} or \code{list}, so that it can be written into HDF5 file.
}
\details{
In this package, \code{h5Prep} will return \code{x} itself by default. 
Extended methods can be easily added for specific S4 class.
}
