% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{hdf5r.Extra-package}
\alias{hdf5r.Extra}
\alias{hdf5r.Extra-package}
\title{hdf5r.Extra: Extensions for 'HDF5' R Interfaces}
\description{
Some methods to manipulate 'HDF5' files, extending the 'hdf5r' package. Reading and writing R objects to 'HDF5' formats follow the specification of 'AnnData' \url{https://anndata.readthedocs.io/en/latest/fileformat-prose.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ycli1995/hdf5r.Extra}
  \item Report bugs at \url{https://github.com/ycli1995/hdf5r.Extra/issues}
}

}
\author{
\strong{Maintainer}: Yuchen Li \email{ycli1995@outlook.com}

}
\keyword{internal}
